fix circular ref
