+++ cocos/3d/builtin/init.ts
@@ -3,9 +3,9 @@ import { ImageAsset } from '../../assets/image-asset';
 import { SpriteFrame } from '../../assets/sprite-frame';
 import { Texture2D } from '../../assets/texture-2d';
 import { GFXDevice } from '../../gfx/device';
+import { selectJointsMediumType } from '../../renderer/models/joints-texture-utils';
 import { TextureCube } from '../assets/texture-cube';
 import effects from './effects';
-import { selectJointsMediumType } from '../../renderer';
 
 class BuiltinResMgr {
     protected _device: GFXDevice | null = null;
